Move models to separate pkg
