+++ src/main/java/hudson/plugins/jira/VersionReleaser.java
@@ -4,6 +4,7 @@ import hudson.model.BuildListener;
 import hudson.model.Result;
 import hudson.model.Run;
 import hudson.model.TaskListener;
+import hudson.plugins.jira.model.JiraVersion;
 
 import java.util.List;
