+++ src/main/java/hudson/plugins/jira/model/JiraVersion.java
@@ -1,4 +1,4 @@
-package hudson.plugins.jira;
+package hudson.plugins.jira.model;
 
 import com.atlassian.jira.rest.client.api.domain.Version;
