+++ src/test/java/hudson/plugins/jira/JiraReleaseVersionUpdaterTest.java
@@ -20,6 +20,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 
+import hudson.plugins.jira.model.JiraVersion;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
