+++ src/test/java/hudson/plugins/jira/JiraVersionCreatorBuilderTest.java
@@ -13,6 +13,7 @@ import java.io.PrintWriter;
 import java.util.HashSet;
 import java.util.Set;
 
+import hudson.plugins.jira.model.JiraVersion;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
