+++ src/test/java/hudson/plugins/jira/deprecated/DeprecatedJiraBuildAction.java
@@ -2,7 +2,7 @@ package hudson.plugins.jira.deprecated;
 
 import hudson.model.AbstractBuild;
 import hudson.model.Action;
-import hudson.plugins.jira.JiraIssue;
+import hudson.plugins.jira.model.JiraIssue;
 import hudson.plugins.jira.Messages;
 
 import java.util.*;
